conditional geoip pull, better kibana resetting logic, re-ordered some steps

+++ supporting-scripts/distro_prep.sh
@@ -66,19 +66,22 @@ rm -f ~elk_user/.bash_history
 echo "cleaning temp directories"
 rm -rf ~elk_user/tmp/*
 
-echo "updating GeoIP database"
+echo "updating GeoIP database.  (Leave both of these blank to skip the GeoIP update.)"
 echo -n "Enter GeoIP AccountID: "
 read geoip_accountid
 echo -n "Enter GeoIP LicenseKey: "
 read geoip_licensekey
-echo "
-AccountID ${geoip_accountid}
-LicenseKey ${geoip_licensekey}
-EditionIDs GeoLite2-Country GeoLite2-City GeoLite2-ASN
-DatabaseDirectory /usr/local/share/GeoIP
-" > ~/GeoIP.conf
-geoipupdate -f ~/GeoIP.conf
-shred -u ~/GeoIP.conf
+
+if [ -z "${geoip_accountid}" -o -z "${geoip_licensekey}" ]; then
+    echo "
+    AccountID ${geoip_accountid}
+    LicenseKey ${geoip_licensekey}
+    EditionIDs GeoLite2-Country GeoLite2-City GeoLite2-ASN
+    DatabaseDirectory /usr/local/share/GeoIP
+    " > ~/GeoIP.conf
+    geoipupdate -f ~/GeoIP.conf
+    shred -u ~/GeoIP.conf
+fi
 
 echo "stopping elastalert"
 systemctl stop elastalert
@@ -90,27 +93,27 @@ curl -s -XDELETE 'http://127.0.0.1:9200/elastalert_status_silence' > /dev/null
 curl -s -XDELETE 'http://127.0.0.1:9200/elastalert_status_status' > /dev/null
 elastalert-create-index --host 127.0.0.1 --port 9200 --no-ssl --no-auth --url-prefix "" --index "elastalert_status" --old-index "" --config /etc/sysconfig/elastalert_config.yml
 
+echo "removing documents from the elasticsearch .kibana index"
+curl -s -H 'kbn-xsrf: true' -H 'Content-Type: application/json' -X POST 'http://localhost:9200/.kibana/_delete_by_query?conflicts=proceed' -d '{"query": { "match_all": {} } }' > /dev/null
+
+echo "reload kibana dashboards"
+/usr/local/sbin/load_all_dashboards.sh
+
 echo "stopping logstash"
-systemctl stop logstash
-echo "removing elasticsearch .kibana index"
-curl -s -XDELETE 'http://localhost:9200/.kibana_1' > /dev/null
-curl -s -XDELETE 'http://localhost:9200/.kibana_2' > /dev/null
-curl -s -XDELETE 'http://localhost:9200/.kibana' > /dev/null
-echo "removing elasticsearch .tasks index"
-curl -s -XDELETE 'http://localhost:9200/.tasks' > /dev/null
+systemctl stop kibana
 
 echo "stopping filebeat service"
 systemctl stop filebeat
 echo "clearing filebeat data"
 if [ -f /var/lib/filebeat/registry ]; then
     echo "filebeat registry is not empty.  The sources below are still tracked.  Press return if this is correct or Ctrl-C to quit."
-    cat /var/lib/filebeat/registry | jq -r .[].source | sed -e 's/^/- /'
+    cat /var/lib/filebeat/registry | jq -r '.[].source' | sed -e 's/^/- /'
     read
 fi
 rm -f /var/lib/filebeat/meta.json
 
-echo "reload kibana dashboards"
-/usr/local/sbin/load_all_dashboards.sh
+echo "removing elasticsearch .tasks index"
+curl -s -XDELETE 'http://localhost:9200/.tasks' > /dev/null
 
 echo "stopping network"
 systemctl stop network
